requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.2.2
dash==2.17.0
plotly==5.22.0
apscheduler==3.10.4
python-dotenv==1.0.1
